refactor: 참여인증 페이지 쿼리옵션 추가 및 챌린지 종료 타이머 위치 이동
